Decouple the orchestrator and validator steps to separate scripts
Each validator will execute the setup-node.sh to initialize its home
folder.
The orchestrator will execute orchestrator-network-setup.sh to aggregate
the inputs from the validators.
The setup-nodes-inverted-flow.sh invokes thea aforementioned scripts in
the background, so that the scripts behave as if they were invoked on
different hosts in parallel.

The introduction is mainly done by these refactoring.

- Move publish_initial_genesis_file() to orchestrator-network-setup.sh
- Move create_home_and_key() to setup-node.sh
- Move each_write_address_to_repo() to setup-node.sh
- Move add_initial_accounts() to orchestrator-network-setup.sh
- Move fetch_genesis_file_with_funds() to setup-node.sh
- Move delegate_stake() to setup-node.sh
- Move sync_apply_gentxs() to setup-node.sh as apply_gentxs()
- Move fetch_genesis_file_with_txs() to setup-node.sh
- Move functions from cosmos-helpers.sh to setup-node.sh
- Make clone_sync_repo() retrying

Signed-off-by: Lars Wegner <lars.wegner2@de.bosch.com>